Modularize project
