+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "bcryptjs",
   "description": "Optimized bcrypt in plain JavaScript with zero dependencies. Compatible to 'bcrypt'.",
-  "version": "2.4.4",
+  "version": "3.0.0",
   "author": "Daniel Wirtz <dcode@dcode.io>",
   "contributors": [
     "Shane Girish <shaneGirish@gmail.com> (https://github.com/shaneGirish)",
